Add ProfilePoints component
