+++ src/components/Profile/ProfilePoints.js
@@ -1,0 +1,12 @@
+import React, { useEffect, useState } from "react";
+
+const ProfilePoints = ({ data, ...props }) => {
+
+
+  return (
+    <section className="profile__main">
+      <h1>Points</h1>
+    </section>)
+}
+
+export default ProfilePoints;
